refactor: multi-file class decomposition
